requirements.txt
asgiref==3.4.1
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
filter==0.0.0.20200724
gunicorn==20.1.0
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
